requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anthropic==0.45.2
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asteroid-filterbanks==0.4.0
attrs==25.1.0
bitsandbytes==0.45.1
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
cut-cross-entropy==25.1.1
cycler==0.12.1
datasets==3.2.0
deepseek==1.0.0
diffusers==0.32.2
dill==0.3.8
distro==1.9.0
docopt==0.6.2
docstring_parser==0.16
einops==0.8.0
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2024.6.1
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
googletrans==4.0.2
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h2==3.2.0
hf_transfer==0.1.9
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.28.1
hyperframe==5.2.0
HyperPyYAML==1.2.2
idna==2.10
importlib_metadata==8.6.1
Jinja2 @ file:///C:/b/abs_10zs3i4lof/croot/jinja2_1737760123028/work
jiter==0.8.2
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.8
lightning==2.5.0.post0
lightning-utilities==0.11.9
loguru==0.7.3
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_a9k8j__c_q/croot/markupsafe_1736370039866/work
matplotlib==3.10.0
mdurl==0.1.2
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.1.0
multiprocess==0.70.16
networkx @ file:///C:/b/abs_b054htfn9t/croot/networkx_1737043671910/work
numpy==1.26.4
omegaconf==2.3.0
openai==1.61.0
optuna==4.2.0
packaging==24.2
pandas==2.2.3
peft==0.14.0
pillow==11.1.0
primePy==1.3
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytz==2024.2
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
shtab==1.7.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.2
SQLAlchemy==2.0.37
srt==3.5.3
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.4.0+cu121
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.4.0+cu121
torchmetrics==1.6.1
tqdm==4.67.1
transformers==4.48.1
triton @ https://github.com/woct0rdho/triton-windows/releases/download/v3.1.0-windows.post5/triton-3.1.0-cp311-cp311-win_amd64.whl#sha256=1628cd027ea4544e73e9827f294f79c6f7186af28debf8dc209f744a74fdbb10
trl==0.14.0
typeguard==4.4.1
typer==0.15.1
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tyro==0.9.13
tzdata==2025.1
unsloth @ git+https://github.com/unslothai/unsloth.git@bdf0cd6033595be4e7ed23d0d002bb176d343152
unsloth_zoo==2025.1.5
uritemplate==4.1.1
urllib3==2.3.0
vosk==0.3.45
websockets==14.2
win32_setctime==1.2.0
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
